trace.c: Optimized structure

+++ src/trace.h
@@ -1,12 +1,10 @@
 #ifndef TRACE_H_
 #define TRACE_H_
 
-#include <stddef.h>
-
 struct Trace {
   char buf[4096];
-  size_t stack[32];
-  size_t stack_size;
+  unsigned int stack[32];
+  unsigned int stack_size;
 };
 typedef struct Trace Trace;
